Convert build.js to build.mjs
