+++ resources/todomvc/architecture-examples/backbone/package.json
@@ -8,7 +8,7 @@
     },
     "private": true,
     "scripts": {
-        "build": "node scripts/build.js",
+        "build": "node scripts/build.mjs",
         "dev": "http-server ./ -p 7001 -c-1 --cors",
         "serve": "http-server ./dist -p 7002 -c-1 --cors"
     },
