+++ resources/todomvc/architecture-examples/jquery-complex/package.json
@@ -2,7 +2,7 @@
   "name": "todomvc-jquery-complex",
   "private": true,
   "scripts": {
-    "build": "node scripts/build.js",
+    "build": "node scripts/build.mjs",
     "serve": "http-server ./dist -p 7002 -c-1 --cors"
   },
   "devDependencies": {
