+++ resources/todomvc/architecture-examples/jquery/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "scripts": {
     "dev": "http-server . -p 7001 -c-1 --cors",
-    "build": "node scripts/build.js",
+    "build": "node scripts/build.mjs",
     "serve": "http-server ./dist -p 7002 -c-1 --cors"
   },
   "dependencies": {
