+++ resources/todomvc/architecture-examples/lit-complex/package.json
@@ -6,7 +6,7 @@
         "npm": ">=8.19.3"
     },
     "scripts": {
-        "build": "node scripts/build.js",
+        "build": "node scripts/build.mjs",
         "serve": "http-server ./dist -p 8000 -c-1 --cors"
     },
     "devDependencies": {
