+++ resources/todomvc/architecture-examples/lit-complex/scripts/build.mjs
@@ -1,25 +1,26 @@
 /**
- * Builds the TodoMVC JavaScript Web Components Complex DOM.
+ * Builds the TodoMVC Lit Complex DOM.
  */
-const path = require("path");
-const { buildComplex } = require("big-dom-generator/utils/buildComplex");
+import path from "path";
+import { buildComplex } from "big-dom-generator/utils/buildComplex.js";
 
 const options = {
     callerDirectory: path.resolve(__dirname),
-    sourceDirectory: path.join("..", "node_modules", "todomvc-javascript-web-components", "dist"),
-    title: "TodoMVC: JavaScript Web Components Complex DOM",
+    sourceDirectory: path.join("..", "node_modules", "todomvc-lit", "dist"),
+    title: "TodoMVC: Lit Complex DOM",
     filesToMove: [
-        "node_modules/big-dom-generator/dist/big-dom.css",
+        "node_modules/big-dom-generator/dist/big-dom-with-stacking-context-scrollable.css",
         "node_modules/big-dom-generator/dist/logo.png",
         "node_modules/big-dom-generator/utils/web-components-css/app.css",
         "node_modules/big-dom-generator/utils/web-components-css/default-variables.css",
-        "node_modules/big-dom-generator/utils/web-components-css/vanilla/todo-item-extra-css.js",
+        "node_modules/big-dom-generator/utils/web-components-css/lit/todo-item-extra-css.js",
         "node_modules/big-dom-generator/utils/web-components-css/todo-list-extra-css.js",
     ],
     extraCssToLink: ["app.css", "default-variables.css"],
     scriptsToLink: ["todo-item-extra-css.js", "todo-list-extra-css.js"],
-    standaloneDirectory: path.resolve(__dirname, "..", "..", "javascript-web-components"),
+    standaloneDirectory: path.resolve(__dirname, "..", "..", "lit"),
     complexDirectory: path.resolve(__dirname, ".."),
+    cssFilesToAddLinksFor: ["big-dom-with-stacking-context-scrollable.css"],
 };
 
 buildComplex(options);
