+++ resources/todomvc/architecture-examples/preact-complex/scripts/build.mjs
@@ -1,15 +1,15 @@
 /**
- * Builds the TodoMVC Svelte Complex DOM.
+ * Build the TodoMVC Preact Complex DOM example.
  */
-const path = require("path");
-const { buildComplex } = require("big-dom-generator/utils/buildComplex");
+import path from "path";
+import { buildComplex } from "big-dom-generator/utils/buildComplex.js";
 
 const options = {
     callerDirectory: path.resolve(__dirname),
-    sourceDirectory: path.join("..", "node_modules", "todomvc-svelte", "dist"),
-    title: "TodoMVC: Svelte Complex DOM",
+    sourceDirectory: path.join("..", "node_modules", "todomvc-preact", "dist"),
+    title: "TodoMVC: Preact Complex DOM",
     filesToMove: ["node_modules/big-dom-generator/dist/big-dom-with-stacking-context-scrollable.css", "node_modules/big-dom-generator/dist/logo.png", "node_modules/big-dom-generator/utils/app.css"],
-    standaloneDirectory: path.resolve(__dirname, "..", "..", "svelte"),
+    standaloneDirectory: path.resolve(__dirname, "..", "..", "preact"),
     complexDirectory: path.resolve(__dirname, ".."),
     cssFilesToAddLinksFor: ["big-dom-with-stacking-context-scrollable.css"],
 };
