+++ resources/todomvc/architecture-examples/vue-complex/package.json
@@ -8,7 +8,7 @@
         "npm": ">=8.19.3"
     },
     "scripts": {
-        "build": "node scripts/build.js",
+        "build": "node scripts/build.mjs",
         "serve": "http-server ./dist -p 7002 -c-1 --cors"
     },
     "devDependencies": {
