+++ resources/todomvc/vanilla-examples/javascript-es5-complex/scripts/build.mjs
@@ -1,17 +1,17 @@
 /**
- * Builds the TodoMVC Backbone Complex DOM.
+ * Builds the TodoMVC JavaScript Es5 Complex DOM.
  */
-const path = require("path");
-const { buildComplex } = require("big-dom-generator/utils/buildComplex");
+import path from "path";
+import { buildComplex } from "big-dom-generator/utils/buildComplex.js";
 
 const options = {
     callerDirectory: path.resolve(__dirname),
-    sourceDirectory: path.join("..", "node_modules", "todomvc-backbone", "dist"),
-    title: "TodoMVC: Backbone Complex DOM",
+    sourceDirectory: path.join("..", "node_modules", "todomvc-javascript-es5", "dist"),
+    title: "TodoMVC: JavaScript Es5 Complex DOM",
     filesToMove: ["node_modules/big-dom-generator/dist/big-dom-with-stacking-context-scrollable.css", "node_modules/big-dom-generator/dist/logo.png"],
     cssFilePath: path.resolve(__dirname, "..", "node_modules", "big-dom-generator", "utils", "app.css"),
     cssFileNamePattern: /^index.*\.css$/,
-    standaloneDirectory: path.resolve(__dirname, "..", "..", "backbone"),
+    standaloneDirectory: path.resolve(__dirname, "..", "..", "javascript-es5"),
     complexDirectory: path.resolve(__dirname, ".."),
     cssFilesToAddLinksFor: ["big-dom-with-stacking-context-scrollable.css"],
 };
