+++ resources/todomvc/vanilla-examples/javascript-web-components-complex/package.json
@@ -8,7 +8,7 @@
     },
     "private": true,
     "scripts": {
-        "build": "node scripts/build.js",
+        "build": "node scripts/build.mjs",
         "serve": "http-server ./dist -p 7006 -c-1 --cors -o"
     },
     "keywords": [],
