+++ resources/todomvc/vanilla-examples/javascript-web-components/package.json
@@ -9,7 +9,7 @@
     "private": true,
     "scripts": {
         "dev": "http-server ./ -p 7005 -c-1 --cors -o",
-        "build": "node scripts/build.js",
+        "build": "node scripts/build.mjs",
         "serve": "http-server ./dist -p 7006 -c-1 --cors -o"
     },
     "keywords": [],
